+++ tap/plugin.py
@@ -2,6 +2,8 @@
 
 import os
 
+from sys import exit
+
 try:
     from unittest import SkipTest
 except ImportError:
@@ -29,7 +31,8 @@ class TAP(Plugin):
             '--tap-format',
             default='',
             help='An optional format string for the TAP output'
-                '{sd} = short_description, {mn} is method name')
+                ' {short_description} is test.shortDescription()'
+                ' {method_name} is str(test)')
 
     def configure(self, options, conf):
         super(TAP, self).configure(options, conf)
@@ -62,5 +65,13 @@ class TAP(Plugin):
 
     def _description(self, test):
         if self._format:
-            return self._format.format(sd=test.shortDescription(), mn=str(test))
+            try:
+                return self._format.format(
+                    short_description=test.shortDescription(),
+                    method_name=str(test))
+            except KeyError as e:
+                exit('''Bad format string: {key}
+Replacement options are: \{short_description\} and \{method_name\}'''.format(
+                    key=e[0]))
+
         return test.shortDescription() or str(test)